Credits to Oleksandr Lobunets and contributors for the code base
Note: I decided to create a new project (instead of forking) as I intend to change some fundamental parts or even the direction.
In the long run, the service discovery should be more conistent to the standard and align to existing implementations like Avahi.
Beside that, some important bugs want to be fixed for higher reliability.