Allow clients to know about leaving servers
When server is going away it sends an multicast packet with TTL=0.
Let the clients who are browsing for this service know about this.
They can check for TTL==0 to know if this is a deletion event.
Without this piece of code, when a server goes away, client does not
know..