Added support for TTL expiry of entry. Before the entry is removed, query is made to server. Only when there is no response, then entry is removed after TTL expiry.
Handle TTL==0 message as per standard. Wait for 1 second before purging
the entry.

Added Browsev3() API : This enabled the user of API to know if a service
is updated or deleted. Also enables the caller to query the current
Cache in addition to maintaining the channel for events.

Browsev2 API - a different version of API.
Not needed but including to generate any discussion needed